+++ src/App.jsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import GlobalStyle from '@styles/GlobalStyles';
+import GlobalStyle from '@styles/common/GlobalStyles';
 import LoginPage from '@pages/LoginPage';
 import SignupPage from '@pages/SignupPage';
 import ProfilePage from '@pages/ProfilePage';
 import MainPage from '@pages/MainPage';
-import TravelRecommendations from '@components/TravelRecommendations'; // @components 별칭을 사용
 
 function App() {
   return (
