+++ src/components/auth/KakaoLoginButton.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const KakaoLoginButton = () => {
+  return <button>Login with Kakao</button>;
+};
+
+export default KakaoLoginButton;
