+++ src/components/auth/LoginForm.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const LoginForm = () => {
+  return (
+    <div>LoginForm</div>
+  )
+}
+
+export default LoginForm
