+++ src/components/main/Carousel.jsx
