+++ src/pages/LoginPage.jsx
@@ -1,45 +1,14 @@
-import React, { useState } from 'react';
-import { LoginContainer, LoginForm, Input, Button } from '@styles/Login.styles';
-import axios from '@axios/api';
-import useBearStore from '@zustands/bearStore';
-import KakaoLoginButton from '@components/KakaoLoginButton';
+import React from 'react';
+import { LoginContainer } from '@styles/auth/Login.styles';
+import LoginForm from '@components/auth/LoginForm';
+import KakaoLoginButton from '@components/auth/KakaoLoginButton';
 
 const LoginPage = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const setUser = useBearStore((state) => state.setUser);
-
-  const handleLogin = async (e) => {
-    e.preventDefault();
-    try {
-      const response = await axios.post('/auth/login', { email, password });
-      setUser(response.data.user);
-      console.log('Login successful:', response.data);
-    } catch (error) {
-      console.error('Login failed:', error);
-    }
-  };
-
   return (
     <LoginContainer>
-      <h1>Login Page</h1>
-      <LoginForm onSubmit={handleLogin}>
-        <Input
-          type="email"
-          placeholder="Email"
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-        />
-        <Input
-          type="password"
-          placeholder="Password"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-        />
-        <Button type="submit">Login</Button>
-      </LoginForm>
+      <h1>Login</h1>
+      <LoginForm />
       <KakaoLoginButton />
-      <Button onClick={() => window.location.href = '/signup'}>Go to Signup</Button>
     </LoginContainer>
   );
 };
