+++ src/pages/MainPage.jsx
@@ -1,15 +1,8 @@
 import React from 'react';
-import styled from 'styled-components';
-import Header from '@components/Header';
-import BrandSection from '@components/BrandSection';
-import TravelRecommendations from '@components/TravelRecommendations';
-
-const MainContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-`;
+import { MainContainer } from '@styles/main/MainPage.style'; // 경로 확인
+import Header from '@components/common/Header'; // 경로 확인
+import BrandSection from '@components/main/BrandSection'; // 경로 확인
+import TravelRecommendations from '@components/main/TravelRecommendations'; // 경로 확인
 
 const MainPage = () => {
   return (
